Remove READ-ALL-FROM-STRING
This function turns out to be unnecessary.
First of all, its PACKAGE keyword argument is not used; second, it
is used to assert that the whole contents of the string have been
read, which does not seem justified by its current use; third, it
is used only in one place in the same file.